requirements.txt
discord.py
PyNaCl
jishaku